fix: unwanted div element in table, separated dnd context from list context
